--- src/main/scala/com/linkedin/feathr/offline/config/datasourceprovider/DataSourceProviderRegistry.scala
@@ -1,7 +1,0 @@
-package com.linkedin.feathr.offline.config.datasourceprovider
-
-import com.linkedin.feathr.offline.config.datasourceprovider.provider.DataSourceProvider
-
-trait DataSourceProviderRegistry {
-  def getDataSourceProvider(name: String): DataSourceProvider
-}
