--- src/main/scala/com/linkedin/feathr/offline/config/datasourceprovider/credential/AccessKeyCredential.scala
@@ -1,11 +1,0 @@
-package com.linkedin.feathr.offline.config.datasourceprovider.credential
-
-import com.fasterxml.jackson.module.caseclass.annotation.CaseClassDeserialize
-import com.linkedin.feathr.offline.config.datasourceprovider.SecretStorage
-
-@CaseClassDeserialize()
-case class AccessKeyCredential(var key: String = "") extends Credential {
-  override def init(ss: SecretStorage, path: String, name: String): Unit = {
-    key = ss.getSecret(path, name, "key")
-  }
-}
