--- src/main/scala/com/linkedin/feathr/offline/config/datasourceprovider/provider/DataSourceProvider.scala
@@ -1,9 +1,0 @@
-package com.linkedin.feathr.offline.config.datasourceprovider.provider
-
-import com.linkedin.feathr.common.DateParam
-import com.linkedin.feathr.offline.config.datasourceprovider.credential.Credential
-import org.apache.spark.sql.{DataFrame, SparkSession}
-
-trait DataSourceProvider {
-  def load(ss: SparkSession): DataFrame
-}
